+++ classes/output/job_overview_table.php
@@ -72,8 +72,8 @@ class job_overview_table extends \table_sql {
         ]);
 
         $this->set_sql(
-            'j.jobid, j.userid, j.timecreated, j.timemodified, j.status, j.statusextras, j.retentiontime, j.artifactfilechecksum, '.
-                'f.pathnamehash, f.filesize, u.username',
+            'j.id, j.jobid, j.userid, j.timecreated, j.timemodified, j.status, j.statusextras, j.retentiontime, '.
+                'j.artifactfilechecksum, f.pathnamehash, f.filesize, u.username',
             '{'.ArchiveJob::JOB_TABLE_NAME.'} j '.
                 'JOIN {user} u ON j.userid = u.id '.
                 'LEFT JOIN {files} f ON j.artifactfileid = f.id',
@@ -164,7 +164,22 @@ class job_overview_table extends \table_sql {
 
         // Action: Show details.
         // @codingStandardsIgnoreLine
-        $html .= '<a href="#" id="job-details-'.$values->jobid.'" class="btn btn-primary mx-1" role="button" title="'.get_string('showdetails', 'admin').'" alt="'.get_string('showdetails', 'admin').'"><i class="fa fa-info-circle"></i></a>';
+        $html .= '<a href="#" id="job-details-'.$values->jobid.'" class="btn btn-primary mx-1" role="button" data-toggle="tooltip" data-placement="top" title="'.get_string('showdetails', 'admin').'" alt="'.get_string('showdetails', 'admin').'"><i class="fa fa-info-circle"></i></a>';
+
+        // Action: Show content.
+        if ($values->pathnamehash) {
+            $showcontentsurl = new \moodle_url('', [
+                'id' => optional_param('id', null, PARAM_INT),
+                'mode' => 'archiver',
+                'action' => 'showcontents',
+                'jobid' => $values->id,
+            ]);
+            // @codingStandardsIgnoreLine
+            $html .= '<a href="'.$showcontentsurl.'" class="btn btn-primary mx-1" role="button" data-toggle="tooltip" data-placement="top" title="'.get_string('inspect_archive_contents', 'quiz_archiver').'" alt="'.get_string('inspect_archive_contents', 'quiz_archiver').'"><i class="fa fa-magnifying-glass"></i></a>';
+        } else {
+            // @codingStandardsIgnoreLine
+            $html .= '<a href="#" class="btn btn-outline-primary disabled mx-1" role="button" alt="'.get_string('inspect_archive_contents', 'quiz_archiver').'" disabled aria-disabled="true"><i class="fa fa-magnifying-glass"></i></a>';
+        }
 
         // Action: Download.
         if ($values->pathnamehash) {
@@ -179,10 +194,8 @@ class job_overview_table extends \table_sql {
                 true,
             );
 
-            $downloadtitle = get_string('download').': '.$artifactfile->get_filename().
-                             ' ('.get_string('size').': '.display_size($artifactfile->get_filesize()).')';
             // @codingStandardsIgnoreLine
-            $html .= '<a href="'.$artifacturl.'" target="_blank" class="btn btn-success mx-1" role="button" title="'.$downloadtitle.'" alt="'.$downloadtitle.'"><i class="fa fa-download"></i></a>';
+            $html .= '<a href="'.$artifacturl.'" target="_blank" class="btn btn-success mx-1" role="button" data-toggle="tooltip" data-placement="top" title="'.get_string('download').'" alt="'.get_string('download').'"><i class="fa fa-download"></i></a>';
         } else {
             // @codingStandardsIgnoreLine
             $html .= '<a href="#" target="_blank" class="btn btn-outline-success disabled mx-1" role="button" alt="'.get_string('download').'" disabled aria-disabled="true"><i class="fa fa-download"></i></a>';
@@ -196,7 +209,7 @@ class job_overview_table extends \table_sql {
             'jobid' => $values->jobid,
         ]);
         // @codingStandardsIgnoreLine
-        $html .= '<a href="'.$deleteurl.'" class="btn btn-danger mx-1" role="button" alt="'.get_string('delete', 'moodle').'"><i class="fa fa-times"></i></a>';
+        $html .= '<a href="'.$deleteurl.'" class="btn btn-danger mx-1" role="button" data-toggle="tooltip" data-placement="top" title="'.get_string('delete').'" alt="'.get_string('delete').'"><i class="fa fa-times"></i></a>';
 
         return $html;
     }