+++ report.php
@@ -39,6 +39,7 @@ use quiz_archiver\Report;
 use quiz_archiver\form\archive_quiz_form;
 use quiz_archiver\form\job_delete_form;
 use quiz_archiver\form\job_sign_form;
+use quiz_archiver\output\archive_contents_table;
 use quiz_archiver\output\job_overview_table;
 
 /**
@@ -520,6 +521,11 @@ class quiz_archiver_report extends report_base {
             }
         }
 
+        // Archive contents table.
+        if (optional_param('action', null, PARAM_TEXT) === 'showcontents') {
+            return $this->generate_archive_contents_page();
+        }
+
         // Archive quiz form.
         if (self::quiz_can_be_archived($this->quiz->id)) {
             $archivequizform = new archive_quiz_form(
@@ -573,6 +579,75 @@ class quiz_archiver_report extends report_base {
         }
 
         return null;
+    }
+
+    /**
+     * Generates the HTML for the archive contents page
+     *
+     * @return string HTML of the archive contents page
+     * @throws \core\exception\coding_exception
+     * @throws \core\exception\moodle_exception
+     * @throws coding_exception
+     * @throws dml_exception
+     * @throws moodle_exception
+     */
+    protected function generate_archive_contents_page(): string {
+        global $OUTPUT;
+
+        // Fetch job.
+        try {
+            $jobid = required_param('jobid', PARAM_INT);
+            $job = ArchiveJob::get_by_id($jobid);
+        } catch (\dml_exception $e) {
+            throw new \moodle_exception('job_not_found', 'quiz_archiver', $this->base_url());
+        }
+
+        // Check if job matches context.
+        if ($job->get_courseid() != $this->course->id ||
+            $job->get_cmid() != $this->cm->id ||
+            $job->get_quizid() != $this->quiz->id
+        ) {
+            throw new \moodle_exception('job_not_found_in_context', 'quiz_archiver', $this->base_url());
+        }
+
+        // Build archive contents table.
+        $expandfilenames = optional_param('expandfilenames', false, PARAM_BOOL);
+        $baseurl = $this->base_url();
+        $baseurl->params([
+            'action' => 'showcontents',
+            'jobid' => $jobid,
+            'expandfilenames' => $expandfilenames,
+        ]);
+
+        $contentstbl = new archive_contents_table('archive_contents_table', $jobid, $expandfilenames);
+        $contentstbl->define_baseurl($baseurl);
+
+        ob_start();
+        $contentstbl->out(50, true);
+        $contentstblhtml = ob_get_contents();
+        ob_end_clean();
+
+        // Generate output HTML.
+        return $OUTPUT->render_from_template(
+            'quiz_archiver/archive_contents',
+            $job->get_metadata() + [
+                'archiveContentsTable' => $contentstblhtml,
+                'backurl' => $this->base_url(),
+                'courseurl' => (new moodle_url('/course/view.php', ['id' => $job->get_courseid()]))->out(),
+                'quizurl' => (new moodle_url('/mod/quiz/view.php', ['id' => $job->get_cmid()]))->out(),
+                'userurl' => (new moodle_url('/user/profile.php', ['id' => $job->get_userid()]))->out(),
+                'attachmentstyleurl' => (new moodle_url($this->base_url(), [
+                    'action' => 'showcontents',
+                    'jobid' => $jobid,
+                    'expandfilenames' => (int) !$expandfilenames,
+                    'page' => optional_param('page', 0, PARAM_INT),
+                ]))->out(),
+                'attachmentstyletext' => get_string(
+                    'attechmentscolumnstyle_'.($expandfilenames ? 'button' : 'list'),
+                    'quiz_archiver'
+                ),
+            ]
+        );
     }
 
     /**